WIP: ProtectedCache Moved for cleaner code and State variable updates for Image Loading indicator
