+++ Photoview/Views/ProtectedImageView.swift
@@ -8,59 +8,6 @@
 import SwiftUI
 import KeychainSwift
 
-protocol ProtectedCache {
-    associatedtype CacheData: AnyObject
-    var cache: NSCache<NSString, CacheData> { get }
-    func parseData(_ data: Data) -> CacheData?
-}
-
-extension ProtectedCache {
-    
-    func fetch(url: String, callback: @escaping (_ data: CacheData) -> Void) -> URLSessionTask? {
-        if let cachedImage = self.cache.object(forKey: url as NSString) {
-            callback(cachedImage)
-            return nil
-        }
-        
-        let request = Network.shared.protectedURLRequest(url: url)
-        
-        let task = URLSession.shared.dataTask(with: request) { data, response, error in
-            if let error = error {
-                if error.localizedDescription != "cancelled" {
-                    print("Error fetching protected image: \(error)")
-                }
-                return
-            }
-            
-            if let data = data, let cacheData = self.parseData(data) {
-                self.cache.setObject(cacheData, forKey: url as NSString)
-                DispatchQueue.main.async {
-                    callback(cacheData)
-                }
-            }
-        }
-        
-        task.resume()
-        
-        return task
-    }
-    
-    func clearCache() {
-        cache.removeAllObjects()
-    }
-    
-}
-
-class ProtectedImageCache: ProtectedCache {
-    static let shared: ProtectedImageCache = ProtectedImageCache()
-    
-    internal let cache: NSCache<NSString, UIImage> = NSCache()
-    
-    func parseData(_ data: Data) -> UIImage? {
-        UIImage(data: data)
-    }
-}
-
 /// Shows a media thumbnail, used in grids where a lot of media is shown at once
 struct ProtectedImageView: View {
     
@@ -68,14 +15,22 @@ struct ProtectedImageView: View {
     let blurhash: String?
     let imageView: (_ image: UIImage) -> AnyView
     
-    init(url: String?, blurhash: String? = nil, imageView: @escaping (_ image: UIImage) -> AnyView) {
+    @Binding var isLoading: Bool
+    
+    init(url: String?, isLoading: Binding<Bool>?, blurhash: String? = nil, imageView: @escaping (_ image: UIImage) -> AnyView) {
         self.url = url
         self.blurhash = blurhash
         self.imageView = imageView
+        if isLoading?.wrappedValue != nil {
+            self._isLoading = isLoading!
+        }
+        else {
+            self._isLoading = .constant(false)
+        }
     }
     
     init(url: String?) {
-        self.init(url: url) { img in
+        self.init(url: url, isLoading: nil) { img in
             AnyView(Image(uiImage: img))
         }
     }
@@ -85,11 +40,13 @@ struct ProtectedImageView: View {
     @State var canceled: Bool = false
     @State var imageLoaded: Bool = false
     
+    @State var numOfRequests: Int = 0
+        
     var presentImage: UIImage? {
         if self.image != nil {
             return self.image
         }
-        
+
         if let blurhash = self.blurhash {
             return UIImage(blurHash: blurhash, size: CGSize(width: 4, height: 3))
         }
@@ -98,7 +55,9 @@ struct ProtectedImageView: View {
     }
     
     func fetchImage(url: String) {
+        numOfRequests += 1
         self.task = ProtectedImageCache.shared.fetch(url: url) { image in
+            numOfRequests -= 1
             self.image = image
         }
     }
@@ -119,6 +78,9 @@ struct ProtectedImageView: View {
                 self.image = nil
             }
         }
+        .onChange(of: numOfRequests) { num in
+            isLoading = num > 0
+        }
         .onAppear {
             if image == nil, let url = url {
                 canceled = false
@@ -127,6 +89,7 @@ struct ProtectedImageView: View {
         }
         .onDisappear {
             canceled = true
+            numOfRequests = 0
             self.task?.cancel()
         }
     }
